refactor: SpaceTest Space역할에 맞는 테스트만으로 구성
